Correcao do tipos de id e criacao da classe PontuacaoBoato

+++ NivelBoato/grails-app/domain/entidade/Boato.groovy
@@ -10,7 +10,7 @@ import java.util.Date
  * @author leonardo
  */
 class Boato {
-    Integer id
+    Long id
     String titulo
     String descricao
     Date dataInicio
